create helper file for controls reducer
